refactor(config_flow): Simplify config flow and fix handler loading
Refactored the config flow to eliminate dynamic imports and merge the `ConfigFlowHandler` logic directly into `config_flow.py`. This resolves the "Flow handler not found" error by ensuring Home Assistant can statically discover the config flow class. The redundant `config_flow_handler.py` file has been removed.
