--- custom_components/meraki_ha/config_flow_handler.py
@@ -1,217 +1,0 @@
-"""Meraki config flow handler."""
-
-import importlib
-import logging
-from typing import Any
-
-import voluptuous as vol
-from homeassistant.config_entries import (
-    AbortFlow,
-    ConfigEntry,
-    ConfigFlow,
-    ConfigFlowResult,
-    OptionsFlow,
-)
-from homeassistant.core import callback
-
-from .const import (
-    CONF_INTEGRATION_TITLE,
-    CONF_MERAKI_API_KEY,
-    CONF_MERAKI_ORG_ID,
-)
-from .schemas import CONFIG_SCHEMA, OPTIONS_SCHEMA
-
-_LOGGER = logging.getLogger(__name__)
-
-
-class ConfigFlowHandler(ConfigFlow):
-    """Handle a config flow for Meraki."""
-
-    VERSION = 1
-    CONNECTION_CLASS = "cloud_poll"
-
-    def __init__(self) -> None:
-        """Initialize the config flow."""
-        self.data: dict[str, Any] = {}
-        self.options: dict[str, Any] = {}
-
-    async def async_step_user(
-        self,
-        user_input: dict[str, Any] | None = None,
-    ) -> ConfigFlowResult:
-        """
-        Handle the initial step.
-
-        Args:
-        ----
-            user_input: The user input.
-
-        Returns
-        -------
-            The flow result.
-
-        """
-        # Dynamically import only the modules we need at runtime to avoid
-        # blocking the event loop during initial import.
-        authentication = importlib.import_module(
-            ".authentication", "custom_components.meraki_ha"
-        )
-        errors_module = importlib.import_module(
-            ".core.errors", "custom_components.meraki_ha"
-        )
-
-        MerakiAuthenticationError = errors_module.MerakiAuthenticationError
-        MerakiConnectionError = errors_module.MerakiConnectionError
-        validate_meraki_credentials = authentication.validate_meraki_credentials
-
-        errors: dict[str, str] = {}
-        if user_input is not None:
-            try:
-                validation_result = await validate_meraki_credentials(
-                    self.hass,
-                    user_input[CONF_MERAKI_API_KEY],
-                    user_input[CONF_MERAKI_ORG_ID],
-                )
-                self.data[CONF_MERAKI_API_KEY] = user_input[CONF_MERAKI_API_KEY]
-                self.data[CONF_MERAKI_ORG_ID] = user_input[CONF_MERAKI_ORG_ID]
-                self.data["org_name"] = validation_result.get(
-                    "org_name",
-                    user_input[CONF_MERAKI_ORG_ID],
-                )
-
-                await self.async_set_unique_id(user_input[CONF_MERAKI_ORG_ID])
-                self._abort_if_unique_id_configured()
-
-                # Show the general form by default
-                return await self.async_step_init()
-
-            except MerakiAuthenticationError:
-                errors["base"] = "invalid_auth"
-            except MerakiConnectionError:
-                errors["base"] = "cannot_connect"
-            except AbortFlow as e:
-                raise e
-            except Exception:
-                _LOGGER.exception("Unexpected exception")
-                errors["base"] = "unknown"
-
-        return self.async_show_form(
-            step_id="user",
-            data_schema=CONFIG_SCHEMA,
-            errors=errors,
-        )
-
-    async def async_step_init(
-        self,
-        user_input: dict[str, Any] | None = None,
-    ) -> ConfigFlowResult:
-        """
-        Handle the general settings step.
-
-        Args:
-        ----
-            user_input: The user input.
-
-        Returns
-        -------
-            The flow result.
-
-        """
-        if user_input is not None:
-            self.options.update(user_input)
-            return self.async_create_entry(
-                title=self.data.get("org_name", CONF_INTEGRATION_TITLE),
-                data=self.data,
-                options=self.options,
-            )
-
-        return self.async_show_form(
-            step_id="init",
-            data_schema=OPTIONS_SCHEMA,
-        )
-
-    @staticmethod
-    @callback
-    def async_get_options_flow(
-        config_entry: ConfigEntry,
-    ) -> OptionsFlow:
-        """
-        Get the options flow for this handler.
-
-        Args:
-        ----
-            config_entry: The config entry.
-
-        Returns
-        -------
-            The options flow handler.
-
-        """
-        # Dynamically import options flow handler
-        options_flow_handler_module = importlib.import_module(
-            ".options_flow", "custom_components.meraki_ha"
-        )
-        return options_flow_handler_module.MerakiOptionsFlowHandler(config_entry)
-
-    async def async_step_reconfigure(
-        self,
-        user_input: dict[str, Any] | None = None,
-    ) -> ConfigFlowResult:
-        """
-        Handle a reconfiguration flow.
-
-        Args:
-        ----
-            user_input: The user input.
-
-        Returns
-        -------
-            The flow result.
-
-        """
-        entry = self.hass.config_entries.async_get_entry(self.context["entry_id"])
-        if not entry:
-            return self.async_abort(reason="unknown_entry")
-
-        if user_input is not None:
-            new_options = {**entry.options, **user_input}
-            self.hass.config_entries.async_update_entry(entry, options=new_options)
-            await self.hass.config_entries.async_reload(entry.entry_id)
-            return self.async_abort(reason="reconfigure_successful")
-
-        schema_with_defaults = self._populate_schema_defaults(
-            OPTIONS_SCHEMA,
-            entry.options,
-        )
-
-        return self.async_show_form(
-            step_id="reconfigure",
-            data_schema=schema_with_defaults,
-        )
-
-    def _populate_schema_defaults(
-        self,
-        schema: vol.Schema,
-        defaults: dict[str, Any],
-    ) -> vol.Schema:
-        """
-        Populate a schema with default values from a dictionary.
-
-        Args:
-        ----
-            schema: The schema to populate.
-            defaults: The default values.
-
-        Returns
-        -------
-            The populated schema.
-
-        """
-        new_schema_keys = {}
-        for key, value in schema.schema.items():
-            if key.schema in defaults:
-                new_key = type(key)(key.schema, default=defaults[key.schema])
-                new_schema_keys[new_key] = value
-            else:
-                new_schema_keys[key] = value
-        return vol.Schema(new_schema_keys)
